==> 3. Análise de conteúdo dos projetos com STM.Rmd <==
---
title: "3. Análise de conteúdo dos projetos com STM"
output: html_notebook
---

# 3. Análise de conteúdo dos projetos com Modelagem de Tópicos

```{r}
#install.packages("dplyr")
#install.packages("data.table")
#install.packages("stm")
#install.packages("tm")
#install.packages("SnowballC")
#install.packages("reshape2")
```

```{r, warning = FALSE}
library(dplyr)
library(here)
library(readr)
library(tidyr)
library(data.table)
library(stringr)
library(tidytext)
library(ggplot2)
```

## 3.1. Preparando a base de dados

```{r}
df <- readRDS(here("Dados/df_agendaclima_municipal.RDS"))


paletas <- readRDS("Dados/paletas.RDS")
plt_proposituras <- paletas$proposituras
```

```{r, 'adicionando_quartis'}
sum_distancia <- summary(df$dias_desastre_anterior)

quartil_1 <- sum_distancia[2]
quartil_2 <- sum_distancia[3]
quartil_3 <- sum_distancia[5]

sum_distancia
```

```{r}


df %>% 
  mutate(Tipo = ifelse(Tipo %in% plt_proposituras$Tipo, Tipo, "outros")) %>% 
  ggplot(aes(x = dias_desastre_anterior, fill = Tipo))+
  geom_histogram(bins=200)+
  geom_vline(xintercept = quartil_1, linetype = "dashed")+
  geom_vline(xintercept = quartil_2, linetype = "dashed")+
  geom_vline(xintercept = quartil_3, linetype = "dashed")+
  theme_linedraw()+
  theme(legend.position = "bottom")+
  scale_color_manual(values = plt_proposituras$Cores, limits = plt_proposituras$Tipo)+
  xlim(0,2000)+
  xlab("Dias desde o desastre anterior")+
  ylab("Número de proposições")
```

```{r}
df <- df %>% 
  mutate(Quartil = case_when(dias_desastre_anterior <= quartil_1 ~ 1,
                             dias_desastre_anterior > quartil_1 & dias_desastre_anterior <= quartil_2 ~ 2,
                             dias_desastre_anterior > quartil_2 & dias_desastre_anterior <= quartil_3 ~ 3,
                             dias_desastre_anterior > quartil_3 ~ 4,
                             .default = NA),
        ementa = as.factor(ementa))

```

## 3.2. Limpando texto das ementas

Antes de analisar o texto das ementas, precisamos limpá-lo para excluir conectivos, pronomes, vocativos e nomes próprios, entre outras palavras que não interessam à análise.

**Certifique-se que o arquivo "nomes_proprios_br.csv" está na pasta "Dados"**

```{r,  warning=FALSE, 'listas_palavras_removidas'}

#Limpando conectivos:
palavras_remover <- data.frame(palavras_remover = c('de',  'a',  'o',  'que',  'e',  'do',  'da',  'em',  'um',  'para',  'é',  'com',  'não',  'uma',  'os',  'no',  'se',  'na',  'por',  'mais',  'as',  'dos',  'como',  'mas',  'foi',  'ao',  'ele',  'das',  'tem',  'à',  'seu',  'sua',  'ou',  'ser',  'quando',  'muito',  'há',  'nos',  'já',  'está',  'eu',  'também',  'só',  'pelo',  'pela',  'até',  'isso',  'ela',  'entre',  'era',  'depois',  'sem',  'mesmo',  'aos',  'ter',  'seus',  'quem',  'nas',  'me',  'esse',  'eles',  'estão',  'você',  'tinha',  'foram',  'essa',  'num',  'nem',  'suas',  'meu',  'às',  'minha',  'têm',  'numa',  'pelos',  'elas',  'havia',  'seja',  'qual',  'será',  'nós',  'tenho',  'lhe',  'deles',  'essas',  'esses',  'pelas',  'este',  'fosse',  'dele',  'tu',  'te',  'vocês',  'vos',  'lhes',  'meus',  'minhas',  'teu',  'tua',  'teus',  'tuas',  'nosso',  'nossa',  'nossos',  'nossas',  'dela',  'delas',  'esta',  'estes',  'estas',  'aquele',  'aquela',  'aqueles',  'aquelas',  'isto',  'aquilo',  'estou',  'está',  'estamos',  'estão',  'estive',  'esteve',  'estivemos',  'estiveram',  'estava',  'estávamos',  'estavam',  'estivera',  'estivéramos',  'esteja',  'estejamos',  'estejam',  'estivesse',  'estivéssemos',  'estivessem',  'estiver',  'estivermos',  'estiverem',  'hei',  'há',  'havemos',  'hão',  'houve',  'houvemos',  'houveram',  'houvera',  'houvéramos',  'haja',  'hajamos',  'hajam',  'houvesse',  'houvéssemos',  'houvessem',  'houver',  'houvermos',  'houverem',  'houverei',  'houverá',  'houveremos',  'houverão',  'houveria',  'houveríamos',  'houveriam',  'sou',  'somos',  'são',  'era',  'éramos',  'eram',  'fui',  'foi',  'fomos',  'foram',  'fora',  'fôramos',  'seja',  'sejamos',  'sejam',  'fosse',  'fôssemos',  'fossem',  'for',  'formos',  'forem',  'serei',  'será',  'seremos',  'serão',  'seria',  'seríamos',  'seriam',  'tenho',  'tem',  'temos',  'tém',  'tinha',  'tínhamos',  'tinham',  'tive',  'teve',  'tivemos',  'tiveram',  'tivera',  'tivéramos',  'tenha',  'tenhamos',  'tenham',  'tivesse',  'tivéssemos',  'tivessem',  'tiver',  'tivermos',  'tiverem',  'terei',  'terá',  'teremos',  'terão',  'teria',  'teríamos',  'teriam', "municipal", 'município', 'municipio', 'municípios', 'municipios', 'prefeito', 'prefeita', 'prefeitos', 'prefeitas', 'executivo', 'secretaria', 'secretarias', 'senhor', 'senhora', 'vereador', 'vereadora', 'vereadores', 'vereadoras', 'poder', 'sr', 'secretário', 'secretária', 'secretários', 'secretárias', 'excelentíssimo', 'excelentíssima', 'excelentíssimos', 'excelentíssimas', 'exmo', 'exma', 'prefeitura', 'prefeituras', 'excelência', 'vossa', 'presidente', 'presidentes', "secretarias", "nobres", 'ministério', 'plenário', 'josé', 'silva', 'joão', 'secadora', 'onety', 'elcione')) %>%  bind_rows((read_csv2(here("Dados/nomes_proprios_br.csv"), col_names = "palavras_remover")))

palavras_remover <- palavras_remover$palavras_remover %>% 
  str_replace_all("([\\^$.|?*+(){}])", "\\\\\\1") %>% 
  str_trim()
  

palavras_remover <- paste0("\\b(", paste(palavras_remover, collapse = "|"), ")\\b")

```

```{r, 'add_ementa_pr'}

df <- df %>%
  mutate(ementa_pr = str_replace_all(ementa, regex(palavras_remover), "_"))
  
rm(palavras_remover)
```

```{r, 'palavras_predominantes_quartil'}

#Cria tabela com palavras, por frequência em que aparecem
palavras_freq <-df %>%
  unnest_tokens(palavras, ementa_pr, drop = F) %>% 
  group_by(palavras) %>% 
  summarise(n = n()) %>% 
  arrange(desc(n)) %>% 
  top_n(30) %>% 
  ggplot(aes(y = reorder(palavras, n), x = n)) +
  geom_col(fill = "black") +
  labs(
    title = "Número de menções aos termos nas ementas",
    x = "Número de menções",
    y = "Palavras",
    caption = "Elaborado pelo autor") +
  theme_classic() +
    theme(legend.position = "none",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 10),
        strip.text = element_text(size = 14))
  
palavras_freq

ggsave(here("Figuras/3.1Frequência das Palavras.png"), palavras_freq)

rm(palavras_freq, tidy)
```

## 3.3 Fazendo análise com STM

```{r, 'preparando_textos'}
library(stm)
########Modelagem de topicos (Structural Topic Model)
#processamento do texto e definição do valor K mais adequado

#Selecionando só textos:
textos <- df$ementa_pr

#Processanto texto (removendo conjunções, separando metadata etc)
prep_texts <- textProcessor(documents=textos, metadata=data.frame(docnumbers=1:length(textos)), language = "portuguese")

#Preparanto o texto: Remove palavras com poucas aparições e ajusta os índices
prep_texts <- prepDocuments(prep_texts$documents, prep_texts$vocab, prep_texts$meta)


textos <- textos[c(prep_texts$meta$docnumbers)]


```

```{r, 'encontrando_k_ideal'}

#Rodar só se necessário (leva muito tempo)
#Valor k (número de tópicos em que dividir)
#K <- 2:30

## searchk procura o melhor número de tópicos
#kresult <- searchK(prep_texts$documents, prep_texts$vocab, K, data=prep_texts$meta)
#plot(kresult)

#saveRDS(kresult, here("Dados/kresult.RDS"))
kresult <- readRDS(here("Dados/kresult.RDS"))
plot(kresult)
#rm(K, kresult)
```

Escolhemos k = 3, 6, 9 para maximizar a Coerência Semântica e 26 para reduzir os resíduos

```{r, 'modelagem_stm'}
#modelagem dos tópicos (9 tópicos)
stm_result <- stm(prep_texts$documents, prep_texts$vocab, 9, data=prep_texts$meta, verbose = F)

saveRDS(stm_result, here("Dados/STM_result.RDS"))

summary(stm_result)
```

```{r, 'thoughts'}

#Localização dos textos mais representativos de determinado tópico
thoughts <-findThoughts(stm_result, texts= textos, topics=c(1:9), n=10)

thoughts_id <- data.frame(thoughts$index)

thoughts_docs <- data.frame(thoughts$docs)

for (i in 1:nrow(thoughts_id)){
  for (j in 1:ncol(thoughts_id)){
    thoughts_docs[i,j] <- as.character(df$ementa[thoughts_id[i,j]]) %>%
      { ifelse(length(.) > 0, ., as.character(NA)) }
  }
}

write.csv(thoughts_docs, here("Dados/STM_thoughts9.csv"))

```

```{r, 'adicionando_ao_df'}
stm_tidy <- tidy(stm_result, matrix = "gamma") %>% 
  pivot_wider(names_from = topic, values_from = gamma, names_prefix = "Tópico ")

df <- bind_cols(df, stm_tidy)%>%
  mutate(tópico_principal = colnames(select(., `Tópico 1`:`Tópico 9`))[max.col(select(., `Tópico 1`:`Tópico 9`), ties.method = "first")])

df <- df[, max_topico := do.call(pmax, c(.SD, na.rm = TRUE)), .SDcols = paste0("Tópico ", 1:9)]

saveRDS(df, file = here("Dados/df_STM.RDS"))



```

```{r}

# Step 1: Flatten thoughts$index into long format
thoughts_long <- data.frame(
  index = unlist(thoughts$index),
  topic = rep(1:9, each = 10) 
)

# Step 2: Subset df by index and bind topic info
df_thought <- df %>%
  mutate(row_id = row_number()) %>%
  inner_join(thoughts_long, by = c("row_id" = "index")) %>%
  arrange(topic)



```

## 3.4 - Cruzando as categorias com os tipos de projeto e desastre

Primeiro, vamos adicionar a paleta de cores para os gráficos com Tópicos:

```{r}
df <- readRDS(here("Dados/df_STM.RDS"))
#Criando paleta de cores para os Tipos: 

paleta_topicos <- data.frame(
  Topicos = c("Tópico 1", "Tópico 2", "Tópico 3", 
              "Tópico 4", "Tópico 5", "Tópico 6", 
              "Tópico 7", "Tópico 8", "Tópico 9"),
  Cores = c("#8E5209", "#09728F", "#8F0909", 
              "#095B8F", "#8F0952", "#09358F", 
              "#098F16", "#091D8F", "#2C098F"))

paletas <- c(paletas, list(topicos = paleta_topicos))

saveRDS(paletas, here("Dados/paletas.RDS"))
```

### 3.4.1 Percentuais atribuídos aos tópicos com maior valor

Depois, vamos analisar quais os percentuais atribuídos aos tópicos com maior percentual

```{r, warning=FALSE}
#Valor médio e variância do tópico com maior percentual de cada projeto
mean_df <- df[, .(mean_val = mean(max_topico, na.rm = TRUE)), by = .(`tópico_principal`)]
mean_df

box_topicos <- df %>% 
  ggplot(aes(y = max_topico, x = tópico_principal, fill = tópico_principal)) +
  geom_boxplot()+
    labs(
      title = "Boxplot: distribuição dos percentuais de Tópicos",
      x = "Tópicos",
      y = "Percentual atribuído aos tópicos com maior valor",
      fill = "Tópicos",
      caption = "Elaborado pelo autor") +
    theme_classic() +
      theme_classic() +
    theme(legend.position = "none",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 10),
        strip.text = element_text(size = 14))+
  ylim(0,1)+
    scale_fill_manual(values = paleta_topicos$Cores, limits = paleta_topicos$Topicos)

box_topicos
  
ggsave(here("Figuras/3.4.1.1Boxplot_Maior_Topico.png"), box_topicos, width = 10, height = 6)
  
rm(box_topicos)

```

Quando mais alta a média e menor a variância no boxplot, maior a certeza na classificação do respectivo Tópico

```{r, 'percent_topicos'}
dist_maior_topic <-  df %>% 
    ggplot(aes(x = max_topico, fill = tópico_principal)) +
    geom_histogram() +
  geom_vline(data = as.data.frame(mean_df), aes(xintercept = mean_val), linetype = "dashed", color = "black") +
  geom_text(data = as.data.frame(mean_df), aes(x = mean_val, y = Inf, label = sprintf("%.2f", mean_val)), vjust = 5,  color = "black", size = 3) +
    labs(
      title = "Distribuição dos percentuais de Tópicos",
      x = "Percentual atribuído aos tópicos",
      y = "Número de proposituras",
      fill = "Tópicos",
      caption = "Elaborado pelo autor") +
    theme_classic() +
        theme(legend.position = "none",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 10),
        strip.text = element_text(size = 14))+
    scale_fill_manual(values = paleta_topicos$Cores, limits = paleta_topicos$Topicos)+
    facet_wrap(~tópico_principal, ncol = 3)

dist_maior_topic

ggsave(here("Figuras/3.4.1.2Histograma_Maior_Topico.png"), dist_maior_topic)
  
rm(dist_maior_topic, mean_df)
```

Interpretamos o gráfico anterior da seguinte forma: quanto mais valores próximos a 1 (direita), menor a incerteza sobre a qual tópico eles pertencem

### 3.4.2 Gráfico Dias desde o último desastre x Tipos

Agora, podemos comparar a soma dos percentuais atribuídos aos projetos em cada dia após o desastre, por tópico

```{r, warning = FALSE, 'graf_dias_tipo'}

topico_dias <- df %>% 
  pivot_longer(starts_with("Tópico "), names_to = "Topicos", values_to = "Percent_Topicos") %>% 
  group_by(dias_desastre_anterior, Topicos) %>% 
  summarise(Percent_Topicos = sum(Percent_Topicos)/n()) %>% 
  filter(dias_desastre_anterior<= (quartil_3 + 365)) %>% 
  ggplot(aes( x = `dias_desastre_anterior`, y = Percent_Topicos, fill = Topicos)) +
  geom_bar(stat = "identity")+
  geom_vline(xintercept = quartil_1, linetype = "dashed")+
  geom_vline(xintercept = quartil_2, linetype = "dashed")+
  geom_vline(xintercept = quartil_3, linetype = "dashed")+
  theme_classic()+
  labs(title = "Percentual atribuído por tópico, por dias após desastre",
       fill = 'Tópico Principal',
       caption = "Elaborado pelo autor")+
  ylab('Soma dos percentuais atribuídos')+
  xlab('Dias desde o desastre anterior')+
      theme(legend.position = "none",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 10),
        strip.text = element_text(size = 14))+
  scale_fill_manual(values = paleta_topicos$Cores, limits = paleta_topicos$Topicos)+
  facet_wrap(~Topicos,  ncol = 3)

topico_dias 


ggsave(here("Figuras/3.4.2.1.Percentual_Topicos_Dias.png"), topico_dias)
  
rm(topico_dias)
```

```{r, 'Por Quartil'}
topico_quartil <- df %>% 
  pivot_longer(starts_with("Tópico "), names_to = "Topicos", values_to = "Percent_Topicos") %>% 
  group_by(Quartil, Topicos) %>% 
  summarise(Percent_Topicos = mean(Percent_Topicos)) %>% 
  ggplot(aes( x = `Quartil`, y = Percent_Topicos, fill = Topicos)) +
  geom_bar(stat = "identity")+
  theme_classic()+
  labs(title = "Percentual atribuído por tópico, por dias após desastre",
       fill = 'Tópico Principal',
       caption = "Elaborado pelo autor")+
  ylab('Soma dos percentuais atribuídos')+
  xlab('Quartis em relação ao desastre anterior')+
        theme(legend.position = "bottom",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 12),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 10),
        strip.text = element_text(size = 14))+
  scale_fill_manual(values = paleta_topicos$Cores, limits = paleta_topicos$Topicos)

topico_quartil

ggsave(here("Figuras/3.4.2.2Percentual_Topicos_Quartil.png"), topico_quartil)
  
rm(topico_quartil)
```

```{r}
topico_quartil <- df %>% 
  group_by(Quartil, tópico_principal) %>% 
  summarise(n = n(),
            Percent_n = n()/sum(n)) %>% 
  ggplot(aes( x = `Quartil`, y = n, fill = tópico_principal)) +
  geom_bar(stat = "identity")+
  theme_classic()+
  labs(title = "Percentual atribuído por tópico, por quartil em relação ao desastre")+
  ylab('Soma dos percentuais atribuídos às proposituras')+
  xlab('Quartis em relação ao desastre anterior')+
  labs(fill = 'Tópico Principal')+
  theme(legend.position = "bottom")+
  scale_fill_manual(values = paleta_topicos$Cores, limits = paleta_topicos$Topicos)

topico_quartil

#ggsave(here("Figuras/3.4.4.Percentual_Topicos_Quartil.png"), topico_quartil)
  
#rm(topico_quartil)
```

```{r}

NDesastres <- readRDS(here("Dados/Desastres.RDS")) %>% 
  filter(Cod_IBGE_Mun %in% df$Cod_IBGE_Mun) %>% 
  group_by(Cod_IBGE_Mun) %>% 
  summarise(n_Desastres = n())

topico_ndesastres <- df %>% 
  pivot_longer(starts_with("Tópico "), names_to = "Topicos", values_to = "Percent_Topicos") %>% 
  group_by(Cod_IBGE_Mun, Topicos) %>% 
  summarise(Percent_Topicos = mean(Percent_Topicos)) %>% 
  merge(NDesastres, by = "Cod_IBGE_Mun", all.x = T) %>% 
  ggplot(aes( x = n_Desastres, y = Percent_Topicos, color = Topicos)) +
  geom_point()+
  theme_classic()+
  labs(title = "Percentual atribuído por tópico, por número de desastres",
       color = 'Tópico Principal',
       caption = "Elaborado pelo autor")+
  ylab('Soma dos percentuais atribuídos')+
  xlab('Número de desastres no município (2013-2024)')+
        theme(legend.position = "none",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 12),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 10),
        strip.text = element_text(size = 14))+
  scale_color_manual(values = paleta_topicos$Cores, limits = paleta_topicos$Topicos)+
  facet_wrap(~Topicos)

topico_ndesastres

ggsave(here("Figuras/3.4.2.3Percentual_Topicos_NDesastres.png"), topico_ndesastres, width = 10, height = 6)
  
rm(topico_ndesastres, NDesastres)
```

### 3.4.3 Por tipo de propositura

```{r, warning=FALSE}
paleta_proposituras <- paletas$proposituras



topico_propositura <- df %>% 
  mutate(Tipo = ifelse(Tipo %in% paleta_proposituras$Tipo, Tipo, "outros")) %>%   
           pivot_longer(starts_with("Tópico "), names_to = "Topicos", values_to = "Percent_Topicos") %>% 
  group_by(Tipo, dias_desastre_anterior, Topicos) %>% 
  summarise(Percent_Topicos = mean(Percent_Topicos))%>% 
  filter(dias_desastre_anterior<= (quartil_3 + 100)) %>% 
  ggplot(aes( x = `dias_desastre_anterior`, y = Percent_Topicos, fill = Topicos)) + 
  geom_bar(stat = "identity")+ 
  geom_vline(xintercept = quartil_1, linetype = "dashed")+
  geom_vline(xintercept = quartil_2, linetype = "dashed")+
  geom_vline(xintercept = quartil_3, linetype = "dashed")+
  theme_bw()+
  labs(title = "Tópicos dos projetos, por dia em relação ao desastre (em percentual)",
       caption = "Elaborado pelo autor")+
  scale_fill_manual(values = paleta_topicos$Cores, limits = paleta_topicos$Topicos)+
  ylab('Percentual de proposituras')+
  xlab('Dias desde o desastre')+
      theme(legend.position = "bottom",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 10),
        strip.text = element_text(size = 14))+
  labs(fill = 'Tópico Principal')+
  xlim(0,600)+
  facet_grid(Topicos~Tipo)

topico_propositura

ggsave(here("Figuras/3.4.3.1Percentual_Topicos_Dias_Tipo_Propositura.png"), topico_propositura, width = 15, height = 10)

rm(topico_propositura)
```

```{r, warning=FALSE}
topico_propositura_quartil <- df %>% 
  mutate(Tipo = ifelse(Tipo %in% paleta_proposituras$Tipo, Tipo, "outros")) %>%   
           pivot_longer(starts_with("Tópico "), names_to = "Topicos", values_to = "Percent_Topicos") %>% 
  group_by(Tipo, Quartil, Topicos) %>% 
  summarise(Percent_Topicos = mean(Percent_Topicos))%>% 
  ggplot(aes( x = `Quartil`, y = Percent_Topicos, fill = Topicos)) + 
  geom_bar(stat = "identity")+ 
  theme_bw()+
  labs(title = "Tópicos dos projetos, por quartil em relação ao desastre (em percentual)",
       caption = "Elaborado pelo autor")+
  scale_fill_manual(values = paleta_topicos$Cores, limits = paleta_topicos$Topicos)+
  ylab('Soma do percentual das proposituras, por tópico')+
  xlab('Quartil em relação ao desastre')+
      theme(legend.position = "bottom",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 12),
        strip.text = element_text(size = 14))+
  labs(fill = 'Tópico Principal')+
  facet_wrap(~Tipo)

topico_propositura_quartil

ggsave(here("Figuras/3.4.3.2Percentual_Topicos_Quartil_Tipo_Propositura.png"), topico_propositura_quartil, width = 10, height = 7)

rm(topico_propositura_quartil)
```

```{r, warning = FALSE}
topico_propositura_quartil <- df %>% 
  mutate(Tipo = ifelse(Tipo %in% paleta_proposituras$Tipo, Tipo, "outros")) %>%  
  group_by(Tipo, Quartil, tópico_principal) %>% 
  summarise(n = n()) %>% 
  mutate(Percent_n = n/sum(n)) %>% 
  ggplot(aes( x = `Quartil`, y = Percent_n, fill = tópico_principal)) + 
  geom_bar(stat = "identity")+ 
  theme_bw()+
  labs(title = "Tópicos dos projetos, por quartil em relação ao desastre (em percentual)",
       caption = "Elaborado pelo autor")+
  scale_fill_manual(values = paleta_topicos$Cores, limits = paleta_topicos$Topicos)+
  ylab('Percentual de proposituras')+
  xlab('Quartil em relação ao desastre')+
      theme(legend.position = "bottom",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 12),
        strip.text = element_text(size = 14))+
  labs(fill = 'Tópico Principal')+
  facet_wrap(~Tipo)

topico_propositura_quartil


rm(topico_propositura_quartil, paleta_proposituras)
```

### 3.4.4 Gráfico Tipos por Desastre

```{r, warning=FALSE}
paleta_desastres <- paletas$desastres

topico_desastres <- df %>%  
  filter(!is.na(desastre_anterior),
         desastre_anterior %in% paleta_desastres$Tipo,
         !(desastre_anterior %in% c('Onda de Calor e Baixa Umidade', 'Onda de Frio', 'Tornado', 'Granizo', NA))) %>% 
           pivot_longer(starts_with("Tópico "), names_to = "Topicos", values_to = "Percent_Topicos") %>% 
  group_by(desastre_anterior, Quartil, Topicos) %>% 
  summarise(Percent_Topicos = mean(Percent_Topicos))%>% 
  ggplot(aes( x = `Quartil`, y = Percent_Topicos, fill = Topicos)) + 
  geom_bar(stat = "identity")+ 
  theme_bw()+
  labs(title = "Tópicos dos projetos, por quartil em relação ao desastre  e tipo de desastre",
       fill = 'Tópico Principal',
       caption = "Elaborado pelo autor")+
  scale_fill_manual(values = paleta_topicos$Cores, limits = paleta_topicos$Topicos)+
  ylab('Percentual de documentos')+
  xlab('Quartil em relação ao desastre')+
  xlab('Quartil em relação ao desastre')+
      theme(legend.position = "none",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 12),
        strip.text = element_text(size = 10))+
  facet_grid(Topicos~desastre_anterior)

topico_desastres

ggsave(here("Figuras/3.4.4.1Percentual_Topicos_Quartil_Desastres_Propositura.png"), topico_desastres, height = 10, width = 15)

rm(topicos_desastres)
```

```{r, warning=FALSE}
topico_desastres <- df %>% 
  filter(!(desastre_anterior %in% c('Onda de Calor e Baixa Umidade', 'Onda de Frio', 'Tornado', 'Granizo', NA))) %>%
  group_by(desastre_anterior, Quartil, tópico_principal) %>% 
  summarise(n = n()) %>%
  group_by(Quartil, desastre_anterior) %>% 
  mutate(porcent_al = n/ sum(n)) %>% 
  ggplot(aes( x = `Quartil`, y = porcent_al, fill = tópico_principal)) + 
  geom_bar(stat = "identity")+ 
  theme_bw()+
  labs(title = "Tópicos dos projetos, por quartil em relação ao desastre  e tipo de desastre")+
  scale_fill_manual(values = paleta_topicos$Cores, limits = paleta_topicos$Topicos)+
  ylab('Percentual de documentos')+
  xlab('Quartil em relação ao desastre')+
  theme(legend.position = "bottom")+
  labs(fill = 'Tópico Principal')+
  facet_grid(tópico_principal~desastre_anterior)

topico_desastres

rm(topicos_desastres)
```

### 3.4.5. Tópicos por ano

```{r, warning = FALSE}
Ano_Topicos <- df %>% 
           pivot_longer(starts_with("Tópico "), names_to = "Topicos", values_to = "Percent_Topicos") %>% 
  group_by(Ano, Quartil, Topicos) %>% 
  summarise(Percent_Topicos = mean(Percent_Topicos))%>% 
  ggplot(aes( x = `Ano`, y = Percent_Topicos, fill = Topicos)) + 
  geom_bar(stat = "identity")+ 
  theme_classic()+
      theme(legend.position = "none",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 12),
        strip.text = element_text(size = 10))+
  labs(title = "Tópicos das proposituras, por Ano",
       fill = 'Tópico',
       caption = "Elaborado pelo autor")+
  scale_fill_manual(values = paleta_topicos$Cores, limits = paleta_topicos$Topicos)+
  scale_x_continuous(n.breaks = 10)+
  ylab('Soma do Percentual dos proposituras')+
  facet_wrap(~Topicos)

Ano_Topicos


rm(Ano_Topicos)
```

```{r, warning=FALSE, 'topico_ano'}
Ano_Topicos <- df %>% 
  group_by(tópico_principal, Ano) %>% 
  summarise(n_al = n()) %>%
  group_by(Ano) %>% 
  mutate(porcent_al = n_al/ sum(n_al))  %>% 
  ggplot(aes( x = `Ano`, y = porcent_al, fill = tópico_principal)) + 
  geom_bar(stat = "identity")+ 
  theme_classic()+
  labs(title = "Tópicos das proposituras, por Ano",
       fill = 'Tópico Principal', 
       caption = "Elaborado pelo autor")+      
  theme(legend.position = "none",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 12),
        strip.text = element_text(size = 10))+
  scale_fill_manual(values = paleta_topicos$Cores, limits = paleta_topicos$Topicos)+
  scale_x_continuous(n.breaks = 3)+
  ylab('Percentual de proposituras')+
    facet_grid(~tópico_principal)

Ano_Topicos

ggsave(here("Figuras/3.4.5P_Topicos_Ano.png"), Ano_Topicos, width = 10, height = 4)

rm(Ano_Topicos)
```

### 3.4.6 Por mês

```{r}
Mes_Topicos <- df %>% 
  mutate(Mes = month(as.Date(Data))) %>% 
  group_by(tópico_principal, Mes) %>% 
  summarise(n_al = n()) %>%
  group_by(Mes) %>% 
  mutate(porcent_al = n_al/ sum(n_al))  %>% 
  ggplot(aes( x = `Mes`, y = porcent_al, fill = tópico_principal)) + 
  geom_bar(stat = "identity")+ 
  theme_classic()+
  labs(title = "Tópicos das proposituras, por Mês",
       fill = 'Tópico Principal', 
       caption = "Elaborado pelo autor")+      
  theme(legend.position = "none",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 12),
        strip.text = element_text(size = 10))+
  scale_fill_manual(values = paleta_topicos$Cores, limits = paleta_topicos$Topicos)+
  scale_x_continuous(n.breaks = 3)+
  ylab('Percentual de proposituras')+
  xlab('Mês')+
    facet_grid(~tópico_principal)


Mes_Topicos

ggsave(here("Figuras/3.4.6P_Topicos_Mes.png"), Mes_Topicos, width=10, height = 4)

rm(Mes_Topicos)
```

## 3.5 Análise só de projetos colocados para votação

Agora, vamos refazer a análise retirando os documentos de pedidos dos vereadores para os Prefeitos, e focando nas proposituras levadas para votação do plenário

```{r}
df_pl <- df %>% 
  filter(!Tipo %in% c("indicacao", "requerimento", "pedido de providencia", "protocolo", "requerimento de informacoes", "oficio", "resposta", "resposta requerimento", "proposições")) 
```

```{r}

########Modelagem de topicos (Structural Topic Model)
#processamento do texto e definição do valor K mais adequado

#Selecionando só textos:
textos <- df_pl$ementa_pr

#Processanto texto (removendo conjunções, separando metadata etc)
prep_texts <- textProcessor(documents=textos, metadata=data.frame(docnumbers=1:length(textos)), language = "portuguese")

#Preparanto o texto: Remove palavras com poucas aparições e ajusta os índices
prep_texts <- prepDocuments(prep_texts$documents, prep_texts$vocab, prep_texts$meta)


textos <- textos[c(prep_texts$meta$docnumbers)]
```

```{r}
#Rodar só se necessário (leva muito tempo)
#Valor k (número de tópicos em que dividir)
#K <- 2:12

## searchk procura o melhor número de tópicos
#kresult <- searchK(prep_texts$documents, prep_texts$vocab, K, data=prep_texts$meta)

#saveRDS(kresult, here("Dados/kresult_pl.RDS"))
#kresult <- readRDS(here("Dados/kresult_pl.RDS"))
#plot(kresult)

#rm(K, kresult) 
```

Escolhi para este k = 9

```{r}
#modelagem dos tópicos (6 tópicos)
stm_result_pl <- stm(prep_texts$documents, prep_texts$vocab, 9, data=prep_texts$meta, verbose = F)

saveRDS(stm_result_pl, here("Dados/stm_result_pl.RDS"))

summary(stm_result_pl)
```

```{r, 'thoughts_PL'}

#Localização dos textos mais representativos de determinado tópico
thoughts <-findThoughts(stm_result_pl, texts= textos, topics=c(1:9), n=10)

thoughts_id <- data.frame(thoughts$index)

thoughts_docs <- data.frame(thoughts$docs)

for (i in 1:nrow(thoughts_id)){
  for (j in 1:ncol(thoughts_id)){
    thoughts_docs[i,j] <- as.character(df_pl$ementa[thoughts_id[i,j]]) %>%
      { ifelse(length(.) > 0, ., as.character(NA)) }
  }
}

write.csv(thoughts_docs, here("Dados/stm_thoughts_pl9.csv"))

```

```{r, 'adicionando_ao_df_PL'}
stm_tidy <- tidy(stm_result_pl, matrix = "gamma") %>% 
  pivot_wider(names_from = topic, values_from = gamma, names_prefix = "Tópico PL")

df_pl <- bind_cols(df_pl, stm_tidy)%>%
  mutate(topico_principalPL = colnames(select(., `Tópico PL1`:`Tópico PL9`))[max.col(select(., `Tópico PL1`:`Tópico PL9`), ties.method = "first")])

df_pl <- df_pl[, max_topicoPL := do.call(pmax, c(.SD, na.rm = TRUE)), .SDcols = paste0("Tópico PL", 1:9)]

saveRDS(df_pl, file = here("Dados/df_STM_PL.RDS"))

rm(textos, prep_texts, stm_tidy, i, j, thoughts_id, thoughts_docs, thoughts)

```

### 3.5.1 - Cruzando as categorias com os tipos de projeto e desastre

Primeiro, vamos adicionar a paleta de cores para os gráficos com Tópicos:

```{r}
df_pl <- readRDS(here("Dados/df_STM_PL.RDS"))
#Criando paleta de cores para os Tipos: 

plt_PLtopicos <- data.frame(
  Topicos = c("Tópico PL1", "Tópico PL2", "Tópico PL3", 
              "Tópico PL4", "Tópico PL5", "Tópico PL6",
              "Tópico PL7", "Tópico PL8", "Tópico PL9"),
  Cores = c("#872121", "#1FAB39", "#38FFDB", 
            "#7F2382", "#44533D", "#ECD736",
            "#FD4FCC", "#1C3D99", "#ED3737"))

paletas <- c(paletas, list(PLtopicos = plt_PLtopicos))

saveRDS(paletas, here("Dados/paletas.RDS"))
```

```{r}
#Verificando as médias

df_pl %>% 
  group_by(topico_principalPL) %>% 
  summarise(n = n()) %>% 
  mutate(percent = n/sum(n))
```

```{r, warning=FALSE}

topPL_diasN <- df_pl %>% 
  pivot_longer(starts_with("Tópico PL"), names_to = "Topicos", values_to = "Percent_Topicos") %>% 
  group_by(dias_desastre_anterior, Topicos) %>% 
  summarise(Percent_Topicos = sum(Percent_Topicos)/n()) %>% 
  ggplot(aes(x = dias_desastre_anterior, y = Percent_Topicos, fill = Topicos))+
  geom_col()+
  geom_vline(xintercept = quartil_1, linetype = "dashed")+
  geom_vline(xintercept = quartil_2, linetype = "dashed")+
  geom_vline(xintercept = quartil_3, linetype = "dashed")+
  theme_bw()+
    theme(legend.position = "bottom",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 10))+
  scale_fill_manual(values = plt_PLtopicos$Cores, limits = plt_PLtopicos$Topicos)+
  xlim(0,2000)+
  labs(x = "Dias desde o desastre anterior",
       y = "Número de proposições",
       title = "Número de proposições por dia após o desastre mais recente (Tópicos PL)",
       fill= "Tópicos",
       caption = "Elaborado pelo autor")+
  facet_wrap(~Topicos)

topPL_diasN

ggsave(here("Figuras/3.5.1.1Dias_Proposituras_Tópicos_PL.png"), topPL_diasN)

rm(topPL_diasN)
```

### 3.5.2 Percentuais atribuídos aos tópicos com maior valor

Depois, vamos analisar quais os percentuais atribuídos aos tópicos com maior percentual

```{r, warning=FALSE}
#Valor médio e variância do tópico com maior percentual de cada projeto

box_topicos <- df_pl %>% 
  ggplot(aes(y = max_topicoPL, x = topico_principalPL, fill = topico_principalPL)) +
  geom_boxplot()+
    labs(
      title = "Boxplot: distribuição dos percentuais de Tópicos (PL)",
      x = "Tópicos",
      y = "Percentual atribuído",
      fill = "Tópicos",
      caption = "Elaborado pelo autor") +
    theme_classic() +
    theme(legend.position = "none",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 10),
        strip.text = element_text(size = 14))+
  ylim(0,1)+
    scale_fill_manual(values = plt_PLtopicos$Cores, limits = plt_PLtopicos$Topicos)

box_topicos
  
ggsave(here("Figuras/3.5.2.1Boxplot_Maior_Topico_PL.png"), box_topicos, width = 10, height = 4)
  
rm(box_topicos)

```

Quando mais alta a média e menor a variância no boxplot, maior a certeza na classificação do respectivo Tópico

```{r, warning=FALSE, 'percent_topicos_pl'}
mean_df <- df_pl[, .(mean_val = mean(max_topicoPL, na.rm = TRUE)), by = .(`topico_principalPL`)]

dist_maior_topic <-  df_pl %>% 
    ggplot(aes(x = max_topicoPL, fill = topico_principalPL)) +
    geom_histogram() +
  geom_vline(data = as.data.frame(mean_df), aes(xintercept = mean_val), linetype = "dashed", color = "black") +
  geom_text(data = as.data.frame(mean_df), aes(x = mean_val, y = Inf, label = sprintf("%.2f", mean_val)), vjust = 5,  color = "black", size = 3) +
    labs(
      title = "Distribuição dos percentuais de Tópicos (PL)",
      x = "Percentual atribuído aos tópicos",
      y = "Número de proposituras",
      fill = "Tópicos",
      caption = "Elaborado pelo autor") +
    theme_classic() +
    theme(legend.position = "none",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 10),
        strip.text = element_text(size = 14))+
    scale_fill_manual(values = plt_PLtopicos$Cores, limits = plt_PLtopicos$Topicos)+
    facet_wrap(~topico_principalPL, ncol = 3)

dist_maior_topic

ggsave(here("Figuras/3.5.2.2Histograma_Maior_Topico_PL.png"), dist_maior_topic, width = 10, height = 6)
  
rm(dist_maior_topic, mean_df)
```

### 3.5.3 Gráfico Dias desde o último desastre x Tipos

Agora, podemos comparar a soma dos percentuais atribuídos aos projetos em cada dia após o desastre, por tópico

```{r, 'graf_dias_tipo_pl'}

topico_dias <- df_pl %>% 
  pivot_longer(starts_with("Tópico PL"), names_to = "Topicos", values_to = "Percent_Topicos") %>% 
  group_by(dias_desastre_anterior, Topicos) %>% 
  summarise(Percent_Topicos = sum(Percent_Topicos)/n()) %>% 
  filter(dias_desastre_anterior<= (quartil_3 + 365)) %>% 
  ggplot(aes( x = `dias_desastre_anterior`, y = Percent_Topicos, fill = Topicos)) +
  geom_bar(stat = "identity", width = 10)+
  geom_vline(xintercept = quartil_1, linetype = "dashed")+
  geom_vline(xintercept = quartil_2, linetype = "dashed")+
  geom_vline(xintercept = quartil_3, linetype = "dashed")+
  theme_classic()+
  labs(title = "Número de proposituras por tópico e dias após desastre",
       caption = "Elaborado pelo autor.")+
  ylab('Número de proposituras')+
  xlab('Dias desde o desastre anterior')+
  labs(fill = 'Tópicos')+
    theme(legend.position = "botton",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 10),
        strip.text = element_text(size = 14))+
  scale_fill_manual(values = plt_PLtopicos$Cores, limits = plt_PLtopicos$Topicos)+
  facet_wrap(~Topicos,  ncol = 3)

topico_dias 


ggsave(here("Figuras/3.5.3.1.Percentual_Topicos_Dias_PL.png"), topico_dias, width = 10, height = 6)
  
rm(topico_dias)
```

```{r, warning=FALSE, 'Por_Quartil_pl'}
topico_quartil <- df_pl %>% 
  pivot_longer(starts_with("Tópico PL"), names_to = "Topicos", values_to = "Percent_Topicos") %>% 
  group_by(Quartil, Topicos) %>% 
  summarise(Percent_Topicos = sum(Percent_Topicos)/n()) %>% 
  ggplot(aes( x = `Quartil`, y = Percent_Topicos, fill = Topicos)) +
  geom_bar(stat = "identity")+
  theme_classic()+
  labs(title = "Percentual de proposituras por tópico e quartil em relação ao desastre")+
  ylab('Soma dos percentuais atribuídos')+
  xlab('Quartis em relação ao desastre anterior')+
  labs(fill = 'Tópico',
       caption = "Elaborado pelo autor.")+
    theme(legend.position = "none",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 10),
        strip.text = element_text(size = 14))+
  scale_fill_manual(values = plt_PLtopicos$Cores, limits = plt_PLtopicos$Topicos)+
  facet_wrap(~Topicos)

topico_quartil

ggsave(here("Figuras/3.5.3.2N_Topicos_Quartil_PL.png"), topico_quartil, width = 10, height = 4)
  
rm(topico_quartil)
```

### 3.5.4 Por tipo de propositura

```{r, warning=FALSE}

topico_propositura <- df_pl %>%
  group_by(Tipo) %>% 
  mutate(n_Tipo = n()) %>% 
  ungroup() %>% 
  mutate(Tipo = ifelse(n_Tipo > 30, Tipo, "outros")) %>%  
  group_by(dias_desastre_anterior, topico_principalPL, Tipo) %>% 
  summarise(n = n()) %>% 
  filter(dias_desastre_anterior<= (quartil_3 + 100)) %>% 
  ggplot(aes( x = `dias_desastre_anterior`, y = n, fill = topico_principalPL)) + 
  geom_bar(stat = "identity", width = 10)+ 
  geom_vline(xintercept = quartil_1, linetype = "dashed")+
  geom_vline(xintercept = quartil_2, linetype = "dashed")+
  geom_vline(xintercept = quartil_3, linetype = "dashed")+
  theme_bw()+
  labs(title = "Tópicos (PL), por dia em relação ao desastre")+
  scale_fill_manual(values = plt_PLtopicos$Cores, limits = plt_PLtopicos$Topicos)+
  ylab('Número proposituras')+
  xlab('Dias desde o desastre')+
    theme(legend.position = "none",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 10),
        strip.text = element_text(size = 14))+
  labs(fill = 'Tópico Principal',
       caption = "Elaborado pelo autor")+
  facet_grid(topico_principalPL~Tipo)

topico_propositura

ggsave(here("Figuras/3.5.4.1Numero_Topicos_PL_Dias_Tipo_Propositura.png"), topico_propositura, width = 15, height = 10)

rm(topico_propositura)
```

```{r, warning=FALSE}
topico_propositura_quartil <- df_pl %>% 
  group_by(Tipo) %>% 
  mutate(n_Tipo = n()) %>% 
  ungroup() %>% 
  mutate(Tipo = ifelse(n_Tipo > 20, Tipo, "outros")) %>%  
           pivot_longer(starts_with("Tópico PL"), names_to = "Topicos", values_to = "Percent_Topicos") %>% 
  group_by(Tipo, Quartil, Topicos) %>%  
  summarise(Percent_Topicos = sum(Percent_Topicos)/n()) %>% 
  ggplot(aes( x = `Quartil`, y = Percent_Topicos, fill = Topicos)) + 
  geom_bar(stat = "identity")+ 
  theme_bw()+
  labs(title = "Tópicos dos projetos, por quartil em relação ao desastre (em percentual)",
       caption = "Elaborado pelo autor")+
  scale_fill_manual(values = plt_PLtopicos$Cores, limits = plt_PLtopicos$Topicos)+
  ylab('Percentual de proposituras')+
  xlab('Quartil em relação ao desastre')+
    theme(legend.position = "bottom",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 10),
        strip.text = element_text(size = 14))+
  labs(fill = 'Tópico Principal')+
  facet_wrap(~Tipo)

topico_propositura_quartil

ggsave(here("Figuras/3.5.4.2N_Topicos_PL_Quartil_Tipo_Propositura.png"), topico_propositura_quartil, width = 12, height = 8)


```

### 3.5.5 Gráfico Tipos por Desastre

```{r, warning=FALSE}

topico_desastres <-  df_pl %>%
  filter(!(desastre_anterior %in% c('Onda de Calor e Baixa Umidade', 'Onda de Frio', 'Tornado', 'Granizo', NA))) %>%  
           pivot_longer(starts_with("Tópico PL"), names_to = "Topicos", values_to = "Percent_Topicos") %>% 
  group_by(desastre_anterior, Quartil, Topicos) %>% 
  summarise(Percent_Topicos = sum(Percent_Topicos)/n()) %>% 
  ggplot(aes( x = `Quartil`, y = Percent_Topicos, fill = Topicos)) + 
  geom_bar(stat = "identity")+ 
  theme_bw()+
  labs(title = "Tópicos (PL) das proposituras, por quartil em relação ao desastre  e tipo de desastre",
       caption = "Elaborado pelo autor")+
  scale_fill_manual(values = plt_PLtopicos$Cores, limits = plt_PLtopicos$Topicos)+
  ylab('Percentual de documentos')+
  xlab('Quartil em relação ao desastre')+
    theme(legend.position = "none",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 10),
        strip.text = element_text(size = 10))+
  labs(fill = 'Tópico')+
  facet_grid(Topicos~desastre_anterior)

topico_desastres

ggsave(here("Figuras/3.5.5NTopicos_Quartil_Desastres_Propositura.png"), topico_desastres, height = 10, width = 15)

rm(topicos_desastres)
```

### 3.5.6. Tópicos por período

```{r, warning=FALSE}
Ano_Topicos <- df_pl %>% 
  group_by(Ano, topico_principalPL) %>% 
  summarise(n = n()) %>% 
  ggplot(aes( x = `Ano`, y = n, fill = topico_principalPL)) + 
  geom_bar(stat = "identity")+ 
  theme_classic()+
  labs(title = "Tópicos das proposituras, por Ano",
       fill = 'Tópico Principal',
       caption = "Elaborado pelo autor")+
  scale_fill_manual(values = plt_PLtopicos$Cores, limits = plt_PLtopicos$Topicos)+
  scale_x_continuous(n.breaks = 4)+
  ylab('Número de proposituras')+
  facet_wrap(~topico_principalPL)+
    theme(legend.position = "none",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 10),
        strip.text = element_text(size = 14))

Ano_Topicos

ggsave(here("Figuras/3.5.6.1N_TopicosPL_Ano.png"), Ano_Topicos, width = 10, height = 5)

rm(Ano_Topicos)
```

```{r, warning=FALSE}
Mes_Topicos <- df_pl %>% 
  mutate(Mês = month(as.Date(Data))) %>% 
  group_by(Mês, topico_principalPL) %>% 
  summarise(n = n()) %>% 
  ggplot(aes( x = `Mês`, y = n, fill = topico_principalPL)) + 
  geom_bar(stat = "identity")+ 
  theme_classic()+
  labs(title = "Tópicos das proposituras, por Mês",
       fill = 'Tópico Principal',
       caption = "Elaborado pelo autor")+
  scale_fill_manual(values = plt_PLtopicos$Cores, limits = plt_PLtopicos$Topicos)+
  scale_x_continuous(n.breaks = 10)+
  ylab('Número de proposituras')+
  facet_wrap(~topico_principalPL)+
    theme(legend.position = "none",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 10),
        strip.text = element_text(size = 14))

Mes_Topicos

ggsave(here("Figuras/3.5.6.2N_TopicosPL_Mes.png"), Mes_Topicos, width = 10, height = 5)

rm(Ano_Topicos, plt_PLtopicos)
```

## 3.6 Construção da base de dados da "agenda climática estrita"

```{r}
dfE <- df_pl %>% 
  filter(Tipo %in% c("projeto de lei ordinaria", "projeto de lei executivo", "projeto de lei complementar", "anteprojeto de lei"),
         topico_principalPL %in% c("Tópico PL2", "Tópico PL3", "Tópico PL4")) 

#Removendo manualmente projetos que não dizem respeito ao clima
PLs_remover <- c("cidade clima", "loteamento clima", "obrigatoriedade de os ônibus", "monóxido de carbono")

# 1. Vetorize: cria matriz lógica [linha x termo] de quais termos aparecem
match_matrix <- sapply(PLs_remover, function(t) grepl(t, dfE$ementa, ignore.case = TRUE))

# 2. Cria a flag projetoclima
dfE[, projetoclima := rowSums(match_matrix) > 0]

#Limpando
dfE <- dfE %>% 
  filter(projetoclima == FALSE)


#verificando número de cidades na nova seleção
dfE %>% 
  select(Cod_IBGE_Mun) %>% 
  unique()


rm(match_matrix, PLs_remover, df_pl)
```

```{r, warning=FALSE}
#Boxplot número de proposituras
df_mun_n<- dfE %>% 
  group_by(Cod_IBGE_Mun, Nome_Municipio, UF) %>% 
  summarise(n = n())

mun_n_sum <- summary(df_mun_n$n)
mun_n_sum  

mun_n_clima <- df_mun_n %>% 
  ggplot(aes(x = n))+
  geom_boxplot(fill = "darkgreen")+ 
  labs(
    title = "Boxplot do número de projetos climáticos (estrito), por municípios",
    y = "",
    x = "Número de proposituras",
    fill = "Origem", 
       caption='Fonte: Elaboração do autor') +
  theme_bw() +
  theme_classic() +
  theme(legend.position = "bottom",
        text = element_text(family = "Times New Roman"),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 14),
        legend.text = element_text(size = 10))

mun_n_clima

ggsave(here("Figuras/3.6.1EDistribuição_N_Municipios_estrito.png"), mun_n_clima, height = 3, width = 10)

rm(mun_n_sum, mun_n_clima)
```

```{r, warning=FALSE}
hist_dfE <- df %>% 
  select(Cod_IBGE_Mun) %>% 
  unique() %>% 
  merge(df_mun_n, by = c("Cod_IBGE_Mun"), all =T) %>% 
  mutate(n = ifelse(is.na(n), 0, n)) %>% 
  ggplot(aes(x = n))+
  geom_histogram(fill = "darkgreen")+ 
  labs(
    title = "Histograma do número de projetos climáticos (estrito), por número de municípios",
    y = "Número de municípios",
    x = "Número de projetos",
    fill = "Origem", 
       caption='Fonte: Elaboração do autor') +
  scale_x_continuous(breaks = c(0,2,4,6,8,10,12,14,16,18,20,22), minor_breaks = c(1,3,5,7,9,11,13,15,17,19,21))+
  theme_bw() +
    theme(legend.position = "none",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 10),
        strip.text = element_text(size = 14))

hist_dfE

ggsave(here("Figuras/3.6.2Histograma_N_Municipios_estrito.png"), hist_dfE, height = 4, width = 10)

rm(hist_dfE, df_mun_n)

```

### Sazonalidade conjunta (Ano e mês)

```{r}
#Adicionando informações dos termos

dfE <- dfE %>%
  mutate(Mes = month(as.Date(Data))) %>% 
  mutate(Mes = ifelse(Mes < 10, paste0(0, Mes), Mes),
         Ano_Mes = paste0(Ano, "-", Mes)) %>%
 separate(clima_termos, c("clima_1", "clima_2", "clima_3", "clima_4", "clima_5", "clima_6", "clima_7"), sep = ", ", fill = "right") %>% 
  pivot_longer(starts_with("clima_"), names_to = "clima_names", names_prefix = "clima_", values_to = "clima_termos") %>% 
  filter(!is.na(clima_termos)) %>% 
  mutate(Categoria = case_when( #Adicionando categorias
  #Chuvas:
  clima_termos %in% c(" pluvial", " pluviais", " chuva", " fortes chuvas",  " chuvoso") ~ "Chuvas",
  #Alagamentos e Inundações:
  clima_termos %in% c(" alagamento",  " inundação", " inundações", " enxurrada", " enchente") ~ "Alagamentos e Inundações",
  #Movimentos de Massa:
  clima_termos %in% c(" deslizamento", " soterramento", " desabamento", " erosão") ~ "Movimentos de Massa",
  #Secas:
  clima_termos %in% c(" seca", " estiagem", " estiagens" ) ~ "Secas",
  #Clima:
  clima_termos %in% c(" clima", " adaptação climática", " aquecimento global", " mitigação", " eventos climáticos extremos", " descarbonização", " transição energética", " carbono", " efeito estufa") ~ "Clima",
  #Ventos:
  clima_termos == " vento" ~ "Ventos",
  #Incêndios:
  clima_termos == " incêndio" ~ "Incêndios",
    #Defesa Civil
    clima_termos == " defesa civil" ~ "Defesa Civil")) %>% 
  ungroup()
```

```{r, warning=FALSE}
paleta_climatermos <- paletas$climatermos
#Criando base de sazonalidade

sazonalidade_grid <- expand_grid(
  Ano = c(2013:2024),
  Mes = c(1:12)) %>% 
  mutate(Mes = ifelse(Mes < 10, paste0(0, Mes), Mes),
         Ano_Mes = paste0(Ano, "-", Mes)) %>% 
  select(Ano_Mes)


# Passo 1: gerar posições (índices) e labels
n_ticks <- nrow(sazonalidade_grid)
x_vals <- sort(unique(sazonalidade_grid$Ano_Mes))

# Criar vetores com as posições desejadas
vlines_idx <- seq(12.5, n_ticks, by = 12)  # linhas verticais após cada ano (após mês 12)
label_idx  <- seq(1, n_ticks, by = 6)    # mostrar só 1 de cada 4 rótulos

# Gráfico
sazon1 <- dfE %>%
  group_by(Ano_Mes, Categoria) %>% 
  summarise(n = n()) %>% 
  merge(sazonalidade_grid, by= c("Ano_Mes"), all = T) %>%
  mutate(n = ifelse(is.na(n), 0, n)) %>% 
  ggplot(aes(y = n, x = factor(Ano_Mes, levels = sort(unique(Ano_Mes))), fill = Categoria)) +  
  geom_col() +
  geom_vline(xintercept = vlines_idx, linetype = "dashed") +
  scale_x_discrete(breaks = x_vals[label_idx]) +
  labs(
    title = "Número de projetos climáticos (estrito), por mês e tipo de projto",
    x = "Ano e Mês",
    y = "Número de propositura",
    fill = "Categoria de Termos Mencionados",
    caption = "Fonte: Elaboração do autor", size = 10
  ) +
  theme_bw() +
  scale_fill_manual(values = paleta_climatermos$Cores, limits = paleta_climatermos$Categoria)+
  theme(
    legend.position = "bottom",
    text = element_text(family = "Times New Roman"),
    axis.title = element_text(size = 14),
    axis.text = element_text(size = 10),
    legend.text = element_text(size = 10),
    axis.text.x = element_text(angle = 45, hjust = 1))

sazon1

ggsave(here("Figuras/3.6.3ESérie_Temporal_Proposituras.png"), sazon1, height = 5, width = 10)

rm(sazon1)
```

```{r, warning=FALSE}

#Heatmap
heatmapE <- dfE %>%
    filter(!is.na(Cod_IBGE_Mun)) %>% 
    group_by(Ano_Mes, Categoria, Cod_IBGE_Mun) %>% 
  summarise(n = n()) %>% 
  merge(sazonalidade_grid, by= c("Ano_Mes"), all = T)
  mutate(n = ifelse(is.na(n), 0, n)) %>% 
  ggplot(aes(y = Cod_IBGE_Mun, x = Ano_Mes)) +  
  geom_tile(aes(fill = Categoria), size = 4)+
  geom_vline(xintercept = vlines_idx, linetype = "dashed") +
  labs(
    title = "Projetos climáticos (estrito), por mês e município",
    x = "Ano e Mês",
    y = "Código IBGE",
    fill = "Categoria de Termos Mencionados",
    caption = "Fonte: Elaboração do autor", size = 10
  ) +
  scale_x_discrete(breaks = x_vals[label_idx]) +
  theme_bw() +
  theme(
    legend.position = "bottom",
    text = element_text(family = "Times New Roman"),
    axis.title = element_text(size = 8),
    axis.text = element_text(size = 8),
    legend.text = element_text(size = 8),
    axis.text.x = element_text(angle = 45, hjust = 1))+
  scale_fill_manual(values = paleta_climatermos$Cores, limits = paleta_climatermos$Categoria)

heatmapE

ggsave(here("Figuras/3.6.4EHeatmap_Proposituras.png"), heatmapE, height = 14, width = 10)

rm(n_ticks, x_vals, vlines_idx, label_idx, heatmapE, sazonalidade, dfheatmap)

```

### Dias desde o último desastre

```{r, warning=FALSE}
prop_dias_desastres <- dfE %>% 
  filter(!is.na(dias_desastre_anterior),
         dias_desastre_anterior <=600) %>% 
  group_by(dias_desastre_anterior, Categoria, desastre_anterior) %>% 
  summarise(n = n()) %>% 
  ggplot(aes(x = dias_desastre_anterior, y = n, fill = Categoria)) +
  geom_col(width = 2)+
  geom_vline(xintercept = 0)+
  geom_vline(xintercept = quartil_1, linetype = "dashed")+
  geom_vline(xintercept = quartil_2, linetype = "dashed")+
  geom_vline(xintercept = quartil_3, linetype = "dashed")+
  labs(title = "Histograma: apresentação de proposituras climáticas desde o desastre mais recente no município", 
       y = "Número de Proposituras",
       x = "Dias desde o desastre mais recente",
       fill = "Termos mencionados",
       caption = "Elaborado pelo autor")+
  scale_fill_manual(values = paleta_climatermos$Cores, limits = paleta_climatermos$Categoria)+
  theme_bw()+
  xlim(0,600)+
    theme(legend.position = "bottom",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 10))+
  facet_wrap(~desastre_anterior)

prop_dias_desastres

ggsave(here("Figuras/3.6.5E Histograma proposituras desde última desastre.png"), prop_dias_desastres, width = 10, height = 6)

rm(prop_dias_desastres, quartil_1, quartil_2, quartil_3)
```

```{r, warning=FALSE}
tiproj_120_dias <- dfE %>%   
  filter(dias_desastre_anterior <= 120) %>% 
  group_by(Categoria, dias_desastre_anterior) %>% 
  summarise(n_al = n()) %>%
  group_by(dias_desastre_anterior) %>% 
  mutate(Percentual = n_al/ sum(n_al)) %>% 
  ggplot(aes( x = `dias_desastre_anterior`, y = Percentual, fill = Categoria)) + 
  geom_bar(stat = "identity", width = 1.1)+ 
  theme_classic()+
  labs(title = "Distribuição dos tipos de proposituras climáticas, por dias após o desastre", 
       y = 'Percentual de projetos', 
       x = "Dias desde o último desastre",
       caption = "Elaborado pelo autor")+
  scale_fill_manual(values = paleta_climatermos$Cores, limits = paleta_climatermos$Categoria)+
  theme(legend.position = "bottom")+
    theme(legend.position = "bottom",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 10),
        strip.text = element_text(size = 14))

tiproj_120_dias

ggsave(here("Figuras/3.6.6Categorias_proposituras_120_dias_tipos_desastre.png"), tiproj_120_dias, width = 10, height = 4)

#rm(tiproj_120_dias, paleta_climatermos)

```

### Nível de Risco Climático

```{r, warning=FALSE}

library(purrr)
library(broom)
#Número de Proposituras apresentadas x Dimensão de Risco

#Todos os municípios
lista_mun <- df %>% 
  select(Cod_IBGE_Mun, AB2_Segurança_Alimentar_Seca, AB2_Segurança_Alimentar_Chuva,  AB2_Desastres_Hidrológicos_Inundações_enxurradas_e_alagamentos, AB2_Desastres_Hidrológicos_Deslizamento_de_terra) %>% 
  rename("AB2_Seca" = AB2_Segurança_Alimentar_Seca,
         "AB2_Chuva" = AB2_Segurança_Alimentar_Chuva,  
         "AB2_Inundações" = AB2_Desastres_Hidrológicos_Inundações_enxurradas_e_alagamentos,
         "AB2_Deslizamentos de Terra" = AB2_Desastres_Hidrológicos_Deslizamento_de_terra) %>% 
  unique()

#Criando dfrisco

dfrisco <- dfE %>% 
  group_by(Cod_IBGE_Mun) %>% 
  summarise(n_clima_estrito = n()) %>% 
  merge(lista_mun, by = "Cod_IBGE_Mun", all=T) %>% 
  mutate(n_clima_estrito = ifelse(is.na(n_clima_estrito), 0, n_clima_estrito)) %>% 
  pivot_longer(cols = starts_with("AB2_"), names_prefix = "AB2_", names_to = "Risco_Climatico", values_to = "Indice_Risco")

annot_df1 <- dfrisco %>%
  group_by(Risco_Climatico) %>%
  group_split() %>%
  map_dfr(~ {
    model <- lm(n_clima_estrito ~ Indice_Risco, data = .x)
    coefs <- tidy(model)
    slope <- coefs %>% filter(term == "Indice_Risco")
    tibble(
      Risco_Climatico = unique(.x$Risco_Climatico),
      x = max(.x$Indice_Risco, na.rm = TRUE),
      y = max(.x$n_clima_estrito, na.rm = TRUE),
      label = paste0("β = ", round(slope$estimate, 3),
                     ", p = ", format.pval(slope$p.value, digits = 2, eps = 0.001))
    )
  })

prop_dimensao_risco <- dfrisco %>% 
  ggplot(aes(y = n_clima_estrito, x = Indice_Risco))+
  geom_point()+
  stat_smooth(method = "lm",
              formula = y ~ x,
              geom = "smooth")+
  geom_text(
    data = annot_df1,
    aes(x = -Inf, y = Inf, label = label),
    inherit.aes = FALSE,
    hjust = -0.1, vjust = 1.1,
    size = 3.5,
    family = "Times New Roman")+
  theme_bw()+
  labs(title = "Percentual de proposituras apresentadas x Indicador de Dimensão de Risco", 
       x = "Índice de Risco",
       y = "Número de Projetos",
       caption = "Elaborado pelo autor")+
  theme(legend.position = "bottom",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 10),
        strip.text = element_text(size = 10))+
  facet_grid(cols = vars(Risco_Climatico))

prop_dimensao_risco

ggsave(here("Figuras/3.6.7.EProposituras_Dimensao_Risco.png"), prop_dimensao_risco, width = 10, height = 4)

rm(prop_dimensao_risco, lista_mun, dfrisco)
```

### Dias até a próxima eleição

```{r, warning=FALSE}
prop_dias_eleicao <- dfE %>% 
  mutate(dias_proxima_eleicao = -1*dias_proxima_eleicao) %>% 
  ggplot(aes(x = dias_proxima_eleicao, fill = T)) +
  geom_histogram(binwidth = 30)+ #Cada um representa aproximadamente 1 mês
  geom_vline(xintercept = 0)+
  geom_vline(xintercept = -365, linetype = "dashed")+
  geom_vline(xintercept = -730, linetype = "dashed")+
  geom_vline(xintercept = -1095, linetype = "dashed")+
  labs(title = "Histograma: apresentação de projetos climáticos (estrito) em relação aos dias até a próxima eleição", 
       y = "Número de Proposituras",
       x = "Dias até a próxima eleição",
       caption = "Elaborado pelo autor")+
  theme_bw()+
  theme(legend.position = "none",
        text = element_text(family = "Times New Roman"),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 10),
        strip.text = element_text(size = 14))+
  xlim(-1461,0)

prop_dias_eleicao

ggsave(here("Figuras/3.6.8E Histograma proposituras até próxima eleição.png"), prop_dias_eleicao, width = 10, height = 5)

rm(prop_dias_eleicao)
```

### Salvando base junto ao df_ols

```{r}
df_ols <- readRDS(here("Dados/df_ols.RDS"))

dfE <- dfE %>% 
  group_by(Cod_IBGE_Mun, Data) %>% 
  summarise(agenda_estrita = n()) %>% 
  data.table()

setDT(dfE)

df_ols <- merge(df_ols, dfE, by=c("Cod_IBGE_Mun", "Data"), all = T)

df_ols[, agenda_estrita := fifelse(is.na(agenda_estrita), 0, agenda_estrita)]

df_ols %>% 
  group_by(agenda_estrita) %>% 
  summarise(n = n())

saveRDS(df_ols, here("Dados/df_ols.RDS"))
```

```{r}
#Removendo todos os arquivos
rm(list = ls(all.names = TRUE))
```

Prossiga para o arquivo "4. Análise dos efeitos dos ECEs sobre a agenda em políticas climáticas"
